backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic>=1.10.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
click==8.1.7
coloredlogs==15.0.1
cryptography==43.0.3
databases==0.9.0
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
durationpy==0.9
ecdsa==0.19.0
emoji==2.14.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
fastapi>=0.115.0,<0.116.0
ffmpy==0.4.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
gradio>=3.41.0,<4.0.0
gradio_client==0.5.0
greenlet==3.1.1
grpcio==1.67.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numexpr==2.10.1
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==0.28.0
orjson==3.10.10
overrides==7.7.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow>=8.0,<11.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic>=2.7.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pypdf==5.1.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
ruff==0.7.1
safehttpx==0.1.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy>=2.0.7
starlette==0.40.0
stripe==11.1.1
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.1
tomli==2.0.2
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.3
unstructured-client==0.26.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
webencodings==0.5.1
wrapt==1.16.0
yarl==1.16.0
zipp==3.20.2
